Major Overhaul
Did a lot of work, changed a lot of things
Entities and Services now have the ability to be added on creating a new
instance! No more need for AddService/AddEntity for most use.

GUI now works better
-TabPosition can be changed in the control, it will auto move new
Controls to the right if placed on an occupied space.
-Controls are kept in state space now, ControlHandler no longer updates
and draws.
-